Move block to its own file and add event_block::await_events
